+++ src/panels/config/network/ha-config-network-ssdp.ts
@@ -1,0 +1,66 @@
+import "@material/mwc-button/mwc-button";
+import type { CSSResultGroup } from "lit";
+import { css, html, LitElement } from "lit";
+import { customElement, property } from "lit/decorators";
+import "../../../components/ha-button";
+import "../../../components/ha-card";
+import { haStyle } from "../../../resources/styles";
+import type { HomeAssistant } from "../../../types";
+
+@customElement("ha-config-network-ssdp")
+class ConfigNetworkSSDP extends LitElement {
+  @property({ attribute: false }) public hass!: HomeAssistant;
+
+  protected render() {
+    return html`
+      <ha-card
+        outlined
+        header=${this.hass.localize("ui.panel.config.network.discovery.ssdp")}
+      >
+        <div class="card-content">
+          <p>
+            ${this.hass.localize("ui.panel.config.network.discovery.ssdp_info")}
+          </p>
+        </div>
+        <div class="card-actions">
+          <a
+            href="/config/ssdp"
+            aria-label=${this.hass.localize(
+              "ui.panel.config.network.discovery.ssdp_browser"
+            )}
+          >
+            <ha-button>
+              ${this.hass.localize(
+                "ui.panel.config.network.discovery.ssdp_browser"
+              )}
+            </ha-button>
+          </a>
+        </div>
+      </ha-card>
+    `;
+  }
+
+  static get styles(): CSSResultGroup {
+    return [
+      haStyle,
+      css`
+        ha-settings-row {
+          padding: 0;
+        }
+
+        .card-actions {
+          display: flex;
+          flex-direction: row-reverse;
+          justify-content: space-between;
+          align-items: center;
+        }
+      `, // row-reverse so we tab first to "save"
+    ];
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    "ha-config-network-ssdp": ConfigNetworkSSDP;
+  }
+}
