+++ src/panels/config/network/ha-config-section-network.ts
@@ -5,7 +5,8 @@ import { isComponentLoaded } from "../../../common/config/is_component_loaded";
 import "../../../layouts/hass-subpage";
 import type { HomeAssistant, Route } from "../../../types";
 import "./ha-config-network";
-import "./ha-config-network-discovery";
+import "./ha-config-network-ssdp";
+import "./ha-config-network-zeroconf";
 import "./ha-config-url-form";
 import "./supervisor-hostname";
 import "./supervisor-network";
@@ -36,10 +37,15 @@ class HaConfigSectionNetwork extends LitElement {
             : ""}
           <ha-config-url-form .hass=${this.hass}></ha-config-url-form>
           <ha-config-network .hass=${this.hass}></ha-config-network>
-          ${isComponentLoaded(this.hass, "zeroconf")
-            ? html`<ha-config-network-discovery
+          ${isComponentLoaded(this.hass, "ssdp")
+            ? html`<ha-config-network-ssdp
                 .hass=${this.hass}
-              ></ha-config-network-discovery>`
+              ></ha-config-network-ssdp>`
+            : ""}
+          ${isComponentLoaded(this.hass, "zeroconf")
+            ? html`<ha-config-network-zeroconf
+                .hass=${this.hass}
+              ></ha-config-network-zeroconf>`
             : ""}
         </div>
       </hass-subpage>
@@ -56,7 +62,8 @@ class HaConfigSectionNetwork extends LitElement {
     supervisor-network,
     ha-config-url-form,
     ha-config-network,
-    ha-config-network-discovery {
+    ha-config-network-ssdp,
+    ha-config-network-zeroconf {
       display: block;
       margin: 0 auto;
       margin-bottom: 24px;
